% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{anxiety_complete}
\alias{anxiety_complete}
\title{Anxiety Data (Complete)}
\format{
A data frame with 580 rows and 4 variables:

\describe{
\item{\code{subject}}{Participant}
\item{\code{treatment}}{Treatment. Either Placebo or Intervention}
\item{\code{measurement}}{Number of measurement}
\item{\code{anxiety}}{Anxiety score (lower is better)}
}
}
\usage{
anxiety_complete
}
\description{
A fictional complete dataset to exemplify the use of HLM method for clinical
significance.
}
\details{
In a fictional clinical trial, participants were split up to belong to either
a medical placebo ("Placebo") or psychotherapeutic intervention
("Intervention") group.

They underwent outpatient treatment during which they were followed for 5
measurements at which a fictional anxiety score was measured. This anxiety
score may range from 0 - 60.

The functional population (i.e., non-anxious
individuals) can be expected to have a mean score of M = 8 points with a
standard deviation of SD = 4.
}
\keyword{datasets}
