% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_anchor.R
\name{calc_anchor.cs_anchor_group_between}
\alias{calc_anchor.cs_anchor_group_between}
\title{Anchor Calculations for Group Effect Between}
\usage{
\method{calc_anchor}{cs_anchor_group_between}(
  data,
  mid_improvement,
  mid_deterioration,
  reference_group,
  post,
  direction,
  ci_level,
  bayesian,
  prior_scale,
  ...
)
}
\arguments{
\item{reference_group}{String, the reference group to compare all other
groups against}

\item{post}{The measurement at which groups should be compared, typically a
measurement after an intervention.}
}
\description{
Anchor Calculations for Group Effect Between
}
\keyword{internal}
