% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_anchor.R
\name{calc_anchor.cs_anchor_group_within}
\alias{calc_anchor.cs_anchor_group_within}
\title{Anchor Calculations for Group Effect Within}
\usage{
\method{calc_anchor}{cs_anchor_group_within}(
  data,
  mid_improvement,
  mid_deterioration,
  direction,
  ci_level,
  bayesian,
  prior_scale,
  ...
)
}
\arguments{
\item{bayesian}{Logical, if Bayesian tests and uncertainty intervals should
be used, defaults to \code{TRUE}.}

\item{prior_scale}{String or numeric, can be adjusted to change the Bayesian
prior distribution. See the documentation for \code{rscale} in
\code{\link[BayesFactor:ttestBF]{BayesFactor::ttestBF()}} for details}
}
\description{
This is an internal function and should never be called directly.
}
\keyword{internal}
