% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_anchor.R
\name{calc_anchor.cs_anchor_individual_within}
\alias{calc_anchor.cs_anchor_individual_within}
\title{Anchor Calculations for Individual Results}
\usage{
\method{calc_anchor}{cs_anchor_individual_within}(data, mid_improvement, mid_deterioration, direction, ci_level, ...)
}
\arguments{
\item{mid_improvement}{Numeric, change that indicates a clinically
significant improvement}

\item{mid_deterioration}{Numeric, change that indicates a clinically
significant deterioration}

\item{direction}{Which direction is beneficial? Lower = -1, better = 1}

\item{ci_level}{Numeric, desired confidence interval.}
}
\description{
Anchor Calculations for Individual Results
}
\keyword{internal}
