% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_anchor.R
\name{calc_anchor}
\alias{calc_anchor}
\title{Generic to Calculate Anchor-Based Results}
\usage{
calc_anchor(data, ...)
}
\arguments{
\item{data}{A data set object of class \verb{cs_*}}

\item{...}{Additional arguments}
}
\value{
An object of class \verb{cs_anchor_*}
}
\description{
This is an internal generic and should not be called directly. Depending on
the different anchor method requested by the user, the appropriate method is
called. It calculates the change and according clinical significance category
for each participant.
}
\keyword{internal}
