% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cutoff.R
\name{calc_cutoff_from_data}
\alias{calc_cutoff_from_data}
\alias{calc_cutoff_from_data.default}
\alias{calc_cutoff_from_data.cs_ha}
\title{Generic for Statistical Approach}
\usage{
calc_cutoff_from_data(data, ...)

\method{calc_cutoff_from_data}{default}(
  data,
  m_clinical,
  sd_clinical,
  m_functional,
  sd_functional,
  type = "a",
  direction = 1,
  ...
)

\method{calc_cutoff_from_data}{cs_ha}(
  data,
  m_clinical,
  sd_clinical,
  m_functional,
  sd_functional,
  m_post,
  sd_post,
  reliability,
  type = "a",
  direction = 1,
  critical_value = 1.65,
  ...
)
}
\arguments{
\item{data}{A pre-processed wide data frame with at least column \code{id}, \code{pre}
and \code{post} and class \verb{cs_*}}

\item{...}{Additional arguments for the respective cutoff method}

\item{m_clinical}{Mean of clinical population}

\item{sd_clinical}{SD of clinical population}

\item{m_functional}{Mean of functional population}

\item{sd_functional}{SD of functional population}

\item{type}{Cutoff type, available are \code{"a"}, \code{"b"}, and \code{"c"}}

\item{direction}{Beneficial intervention effect for given instrument. 1 =
higher is better, -1 = lower is better}

\item{m_post}{Mean of post measurement}

\item{sd_post}{SD of post measurement}

\item{reliability}{Instrument's reliability}

\item{critical_value}{The critical value for the RCI decision, should be
1.65 if significance_level = 0.05}
}
\value{
A list with cutoff info and participant wise info on cutoff
categorization
}
\description{
This is an internal generic and should not be called directly. Depending on
the different cutoff method requested by the user, the appropriate method is
called. It calculates the cutoff and according clinical significance category
for each participant.
}
