% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cutoff.R
\name{calc_cutoff_from_data.cs_ha}
\alias{calc_cutoff_from_data.cs_ha}
\title{Calculate cs_indiv}
\usage{
\method{calc_cutoff_from_data}{cs_ha}(
  data,
  m_clinical,
  sd_clinical,
  m_functional,
  sd_functional,
  m_post,
  sd_post,
  reliability,
  type = "a",
  direction = 1,
  critical_value = 1.65,
  ...
)
}
\arguments{
\item{m_clinical}{Mean of clinical population}

\item{sd_clinical}{SD of clinical population}

\item{m_post}{Mean of post measurement}

\item{sd_post}{SD of post measurement}

\item{reliability}{Instrument's reliability}

\item{critical_value}{The critical value for the RCI decision, should be
1.65 if significance_level = 0.05}
}
\description{
Calculate cs_indiv
}
\keyword{internal}
