% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cutoff.R
\name{calc_cutoff_from_data.default}
\alias{calc_cutoff_from_data.default}
\title{Calculate the categories based on the cutoff}
\usage{
\method{calc_cutoff_from_data}{default}(
  data,
  m_clinical,
  sd_clinical,
  m_functional,
  sd_functional,
  type = "a",
  direction = 1,
  ...
)
}
\arguments{
\item{data}{Prepped data with class \verb{cs_*}}

\item{m_functional}{Mean of functional population}

\item{sd_functional}{SD of functional population}

\item{type}{Cutoff type, available are \code{"a"}, \code{"b"}, and \code{"c"}}

\item{...}{Additional arguments for the specific RCI method}
}
\description{
Calculate the categories based on the cutoff
}
\keyword{internal}
