% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cutoff.R
\name{calc_cutoff_from_data}
\alias{calc_cutoff_from_data}
\title{Generic for Statistical Approach}
\usage{
calc_cutoff_from_data(data, ...)
}
\arguments{
\item{data}{A pre-processed wide data frame with at least column \code{id}, \code{pre}
and \code{post} and class \verb{cs_*}}

\item{...}{Additional arguments for the respective cutoff method}
}
\value{
A list with cutoff info and participant wise info on cutoff
categorization
}
\description{
This is an internal generic and should not be called directly. Depending on
the different cutoff method requested by the user, the appropriate method is
called. It calculates the cutoff and according clinical significance category
for each participant.
}
\keyword{internal}
