% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_percentage.R
\name{calc_percentage}
\alias{calc_percentage}
\title{Calculate Change for the Percentage-Change Approach}
\usage{
calc_percentage(data, pct_improvement, pct_deterioration, direction)
}
\arguments{
\item{data}{A pre-processed data frame with at least columns \code{change} and
\code{pre}}

\item{pct_improvement}{Numeric, the required percent change cutoff to infer
an improvement}

\item{pct_deterioration}{Numeric, the required percent change cutoff to infer
a deterioration}
}
\value{
An object of classm \code{cs_percentage} list with participant-wise info
on clinical significant change catergory
}
\description{
This is an internal generic and should not be called directly. Depending on
the different percent change cutoff requested by the user, this function
calculates the relative change according clinical significance category for
each patient.
}
\keyword{internal}
