% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rci.R
\name{calc_rci.cs_en}
\alias{calc_rci.cs_en}
\title{RCI for the Edwards method}
\usage{
\method{calc_rci}{cs_en}(
  data,
  m_pre,
  sd_pre,
  reliability,
  direction = 1,
  critical_value = 1.96,
  ...
)
}
\description{
RCI for the Edwards method
}
\keyword{internal}
