% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rci.R
\name{calc_rci.cs_gln}
\alias{calc_rci.cs_gln}
\title{RCI for the Gulliksen, Lord & Novick method}
\usage{
\method{calc_rci}{cs_gln}(
  data,
  m_pre,
  sd_pre,
  reliability,
  direction = 1,
  critical_value = 1.96,
  ...
)
}
\arguments{
\item{m_pre}{Pre measurement mean}
}
\description{
RCI for the Gulliksen, Lord & Novick method
}
\keyword{internal}
