% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rci.R
\name{calc_rci.cs_ha}
\alias{calc_rci.cs_ha}
\title{RCI for the Hageman & Arrindell}
\usage{
\method{calc_rci}{cs_ha}(
  data,
  m_pre,
  sd_pre,
  m_post,
  sd_post,
  reliability,
  direction = 1,
  critical_value = 1.65,
  ...
)
}
\arguments{
\item{sd_post}{Post measurement SD}
}
\description{
RCI for the Hageman & Arrindell
}
\keyword{internal}
