% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rci.R
\name{calc_rci.cs_hll}
\alias{calc_rci.cs_hll}
\title{RCI for the Hsu, Lin & Lord method}
\usage{
\method{calc_rci}{cs_hll}(
  data,
  m_pre,
  sd_pre,
  m_post,
  reliability,
  direction = 1,
  critical_value = 1.96,
  ...
)
}
\arguments{
\item{m_post}{Post measurement mean}
}
\description{
RCI for the Hsu, Lin & Lord method
}
\keyword{internal}
