% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rci.R
\name{calc_rci.cs_hlm}
\alias{calc_rci.cs_hlm}
\title{Calc RCI for the HLM method}
\usage{
\method{calc_rci}{cs_hlm}(data, direction, critical_value = 1.96, ...)
}
\description{
Calc RCI for the HLM method
}
\keyword{internal}
