% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rci.R
\name{calc_rci.cs_nk}
\alias{calc_rci.cs_nk}
\title{RCI for the NK method}
\usage{
\method{calc_rci}{cs_nk}(
  data,
  m_pre,
  sd_pre,
  reliability,
  reliability_post,
  direction = 1,
  critical_value = 1.96,
  ...
)
}
\arguments{
\item{reliability_post}{Instrument reliability at post measurement}
}
\description{
RCI for the NK method
}
\keyword{internal}
