% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_significance.R
\name{clinical_significance}
\alias{clinical_significance}
\title{Clinical Significance}
\usage{
clinical_significance(
  data,
  id,
  time,
  outcome,
  group = NULL,
  pre = NULL,
  post = NULL,
  m_functional = NA,
  sd_functional = NA,
  type = "a",
  reliability,
  reliability_post,
  better_is = c("lower", "higher"),
  method = c("JT", "GLN", "HLL", "EN", "NK", "HA", "HLM")
)
}
\arguments{
\item{data}{A tidy data frame}

\item{id}{Participant ID}

\item{time}{Time variable}

\item{outcome}{Outcome variable}

\item{group}{Grouping variable}

\item{pre}{Pre measurement}

\item{post}{Post measurement}

\item{m_functional}{Mean of the functional population}

\item{sd_functional}{Standard deviation of the functional population}

\item{type}{Cutoff type. Available are \code{"a"}, \code{"b"}, and \code{"c"}. Defaults to
\code{"a"} (see details for further information in which cutoff to choose).}

\item{reliability}{The instrument's reliability estimate. If you selected the
NK method, the here specified reliability will be the instrument's pre
measurement reliability.}

\item{reliability_post}{The instrument's reliability at post measurement.
This is only relevant for the NK method.}

\item{better_is}{Which direction means a better outcome? Available are
\code{"lower"} and \code{"higher"}. Defaults to \code{"lower"}.}

\item{method}{Clinical significance method. Available are
\itemize{
\item \code{"JT"} (Jacobson & Truax)
\item \code{"GLN"} (Gulliksen, Lord, and Novick)
}}
}
\value{
An S3 object of class \code{clinisig}
}
\description{
Clinical Significance
}
\examples{
# Clinical significane for "negative" outcomes (lower values are desirable)
jacobson_1989 \%>\%
  clinical_significance(
    id = subject,
    time = time,
    outcome = gds,
    pre = "pre",
    reliability = 0.80
  )


# Clinical significane for "positive" outcomes (higher values are desirable)
jacobson_1989 \%>\%
  clinical_significance(
    id = subject,
    time = time,
    outcome = das,
    pre = "pre",
    reliability = 0.80,
    better_is = "higher"
  )


# Clinical significance incorporating descriptives of a functional population.
# Make sure to select type = "c" to incorporate the specified functional
# descriptives.
jacobson_1989 \%>\%
  clinical_significance(
    id = subject,
    time = time,
    outcome = gds,
    pre = "pre",
    reliability = 0.80,
    m_functional = 30,
    sd_functional = 7,
    type = "c"
  )


# Change the clinical significance method
jacobson_1989 \%>\%
  clinical_significance(
    id = subject,
    time = time,
    outcome = gds,
    pre = "pre",
    reliability = 0.80,
    m_functional = 30,
    sd_functional = 7,
    type = "c",
    method = "EN"
  )

jacobson_1989 \%>\%
  clinical_significance(
    id = subject,
    time = time,
    outcome = gds,
    pre = "pre",
    reliability = 0.80,
    m_functional = 30,
    sd_functional = 7,
    type = "c",
    method = "HA"
  )


# And plot your results
results <- jacobson_1989 \%>\%
  clinical_significance(
    id = subject,
    time = time,
    outcome = gds,
    pre = "pre",
    reliability = 0.80,
    m_functional = 30,
    sd_functional = 7,
    type = "c"
  )

plot(results)
}
