% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_summary_table.R
\name{create_summary_table}
\alias{create_summary_table}
\alias{create_summary_table.cs_distribution}
\alias{create_summary_table.cs_statistical}
\alias{create_summary_table.cs_combined}
\alias{create_summary_table.cs_percentage}
\alias{create_summary_table.cs_anchor_individual_within}
\title{Generic For Creating Summary Tables}
\usage{
create_summary_table(x, ...)

\method{create_summary_table}{cs_distribution}(x, data, ...)

\method{create_summary_table}{cs_statistical}(x, data, method, ...)

\method{create_summary_table}{cs_combined}(
  x,
  cutoff_results,
  data,
  method,
  r_dd,
  se_measurement,
  cutoff,
  sd_post,
  direction,
  ...
)

\method{create_summary_table}{cs_percentage}(x, data, ...)

\method{create_summary_table}{cs_anchor_individual_within}(x, data, ...)
}
\arguments{
\item{x}{A results object that differs per approach:
\itemize{
\item RCI results for distribution-based and combined approach
\item Cutoff results for statistical approach
\item PCT results for percentage-change approach
\item Anchor results for the anchor-based approach
}}

\item{...}{Additional arguments}

\item{data}{The used data frame}

\item{method}{Clinical significance method}

\item{cutoff_results}{Cutoff results object}

\item{r_dd}{r_dd}

\item{se_measurement}{se_measurement}

\item{cutoff}{Cutoff value}

\item{sd_post}{SD post intervention}

\item{direction}{Which direction is beneficial? 1 = higher, -1 = lower}
}
\value{
A tibble
}
\description{
This is an internal function used to create summary table for the different
clinical significance methods. This should not be used directly.
}
