% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_summary_table.R
\name{create_summary_table.cs_anchor_individual_within}
\alias{create_summary_table.cs_anchor_individual_within}
\title{Create Summary Table for Anchor-Based Approach}
\usage{
\method{create_summary_table}{cs_anchor_individual_within}(x, data, ...)
}
\description{
Create Summary Table for Anchor-Based Approach
}
\keyword{internal}
