% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_summary_table.R
\name{create_summary_table.cs_combined}
\alias{create_summary_table.cs_combined}
\title{Create Summary Table for Combined Approach}
\usage{
\method{create_summary_table}{cs_combined}(
  x,
  cutoff_results,
  data,
  method,
  r_dd,
  se_measurement,
  cutoff,
  sd_post,
  direction,
  ...
)
}
\arguments{
\item{cutoff_results}{Cutoff results object}

\item{r_dd}{r_dd}

\item{se_measurement}{se_measurement}

\item{cutoff}{Cutoff value}

\item{sd_post}{SD post intervention}

\item{direction}{Which direction is beneficial? 1 = higher, -1 = lower}
}
\description{
Create Summary Table for Combined Approach
}
\keyword{internal}
