% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_summary_table.R
\name{create_summary_table.cs_distribution}
\alias{create_summary_table.cs_distribution}
\title{Create Summary Table for Distribution-Based Approach}
\usage{
\method{create_summary_table}{cs_distribution}(x, data, ...)
}
\arguments{
\item{x}{A results object that differs per approach:
\itemize{
\item RCI results for distribution-based and combined approach
\item Cutoff results for statistical approach
\item PCT results for percentage-change approach
\item Anchor results for the anchor-based approach
}}

\item{data}{The used data frame}

\item{...}{Additional arguments}
}
\description{
Create Summary Table for Distribution-Based Approach
}
\keyword{internal}
