% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_summary_table.R
\name{create_summary_table.cs_percentage}
\alias{create_summary_table.cs_percentage}
\title{Create Summary Table for Percentage-Change Approach}
\usage{
\method{create_summary_table}{cs_percentage}(x, data, ...)
}
\description{
Create Summary Table for Percentage-Change Approach
}
\keyword{internal}
