% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_summary_table.R
\name{create_summary_table.cs_statistical}
\alias{create_summary_table.cs_statistical}
\title{Create Summary Table for Statistical Approach}
\usage{
\method{create_summary_table}{cs_statistical}(x, data, method, ...)
}
\arguments{
\item{method}{Clinical significance method}
}
\description{
Create Summary Table for Statistical Approach
}
\keyword{internal}
