% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_summary_table.R
\name{create_summary_table}
\alias{create_summary_table}
\title{Generic For Creating Summary Tables}
\usage{
create_summary_table(x, ...)
}
\arguments{
\item{x}{An object of class \verb{cs_*}}

\item{...}{Additional arguments passed to other methods}
}
\value{
A tibble
}
\description{
This is an internal function used to create summary table for the different
clinical significance methods. This should not be used directly.
}
\keyword{internal}
