% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_get_augmented_data.R
\name{cs_get_augmented_data.cs_anchor_individual_within}
\alias{cs_get_augmented_data.cs_anchor_individual_within}
\title{Extract Augmented Data from a cs_anchor_individual Object}
\usage{
\method{cs_get_augmented_data}{cs_anchor_individual_within}(x, ...)
}
\arguments{
\item{x}{An object of class \code{cs_anchor_individual_within}}

\item{...}{Additional arguments}
}
\value{
A tibble
}
\description{
Extract Augmented Data from a cs_anchor_individual Object
}
