% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_get_augmented_data.R
\name{cs_get_augmented_data.cs_combined}
\alias{cs_get_augmented_data.cs_combined}
\title{Extract Augmented Data from a cs_combined Object}
\usage{
\method{cs_get_augmented_data}{cs_combined}(x)
}
\value{
A tibble with used data and clinical significance categories
}
\description{
To obtain patient-wise results.
}
\details{
This function returns the patient-wise results, containing the considered pre
and post intervention value, its raw change as well as the RCI and the
individual category a patient belongs to.
}
\examples{
results <- jacobson_1989 |>
clinical_significance(
  id = subject,
  time = time,
  outcome = gds,
  pre = "pre",
  reliability = 0.80,
  m_functional = 30,
  sd_functional = 7,
  type = "c"
)

cs_get_augmented_data(results)
}
