% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_get_augmented_data.R
\name{cs_get_augmented_data.cs_percentage}
\alias{cs_get_augmented_data.cs_percentage}
\title{Extract Augmented Data from a cs_percentage Object}
\usage{
\method{cs_get_augmented_data}{cs_percentage}(x, ...)
}
\arguments{
\item{x}{An object of class \code{cs_percentage}}

\item{...}{Additional arguments}
}
\value{
A tibble
}
\description{
Extract Augmented Data from a cs_percentage Object
}
