% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_get_augmented_data.R
\name{cs_get_augmented_data.cs_statistical}
\alias{cs_get_augmented_data.cs_statistical}
\title{Extract Augmented Data from a cs_statistical Object}
\usage{
\method{cs_get_augmented_data}{cs_statistical}(x, ...)
}
\arguments{
\item{x}{An object of class \code{cs_distribution}}

\item{...}{Additional arguments}
}
\value{
A tibble
}
\description{
Extract Augmented Data from a cs_statistical Object
}
