% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_get_augmented_data.R
\name{cs_get_augmented_data}
\alias{cs_get_augmented_data}
\title{Generic to Extract Augmented Data from a clinisig Object}
\usage{
cs_get_augmented_data(x, ...)
}
\arguments{
\item{x}{A clinisig object}

\item{...}{Additional arguments}
}
\value{
A tibble with augmented data
}
\description{
Generic to Extract Augmented Data from a clinisig Object
}
