% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_get_cutoff.R
\name{cs_get_cutoff}
\alias{cs_get_cutoff}
\title{Get Used Cutoff And Type From A clinisig Object}
\usage{
cs_get_cutoff(x, with_descriptives = FALSE)
}
\arguments{
\item{x}{A clinisig object}

\item{with_descriptives}{Logical indicating whether you want to retrieve only
the cutoff type and value or the summary statistics on which it is based
on. The default is \code{FALSE}.}
}
\value{
A tibble with cutoff information
}
\description{
Get Used Cutoff And Type From A clinisig Object
}
