% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_get_cutoff_descriptives.R
\name{cs_get_cutoff_descriptives}
\alias{cs_get_cutoff_descriptives}
\title{Get Descriptives Used In The Cutoff Calculation}
\usage{
cs_get_cutoff_descriptives(x)
}
\arguments{
\item{x}{A clinisig object}
}
\value{
A tibble with means and standard deviations of the clinical and
functional population
}
\description{
Get Descriptives Used In The Cutoff Calculation
}
