% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_get_reliability.R
\name{cs_get_reliability}
\alias{cs_get_reliability}
\title{Get Reliability Of A clinisig Object}
\usage{
cs_get_reliability(x)
}
\arguments{
\item{x}{A clinisig object}
}
\value{
A tibble showing the reliability
}
\description{
Get Reliability Of A clinisig Object
}
\examples{
results <- jacobson_1989 |>
  clinical_significance(
    id = subject,
    time = time,
    outcome = gds,
    pre = "pre",
    reliability = 0.80,
    m_functional = 30,
    sd_functional = 10,
    type = "c"
  )

results_nk <- jacobson_1989 |>
  clinical_significance(
    id = subject,
    time = time,
    outcome = gds,
    pre = "pre",
    reliability = 0.80,
    reliability_post = 0.85,
    m_functional = 30,
    sd_functional = 10,
    type = "c",
    method = "NK"
  )

cs_get_reliability(results)
cs_get_reliability(results_nk)
}
\seealso{
Extractor functions
\code{\link{cs_get_augmented_data}()},
\code{\link{cs_get_data}()},
\code{\link{cs_get_n}()},
\code{\link{cs_get_summary}()}
}
\concept{get}
