% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_get_summary.R
\name{cs_get_summary.cs_anchor_group_within}
\alias{cs_get_summary.cs_anchor_group_within}
\title{Method Get Summary Table From A cs_analysis Object With Within Group Effects}
\usage{
\method{cs_get_summary}{cs_anchor_group_within}(x)
}
\arguments{
\item{x}{An object of class \code{cs_analysis}}
}
\description{
Method Get Summary Table From A cs_analysis Object With Within Group Effects
}
