% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_get_summary.R
\name{cs_get_summary.default}
\alias{cs_get_summary.default}
\title{Default Method To Get A Summary Table From A cs_analysis Object}
\usage{
\method{cs_get_summary}{default}(x, which = c("individual", "group"))
}
\arguments{
\item{x}{An object of class \code{cs_analysis}}

\item{which}{Which level of summary table to return. This is only necessary
for method \code{"HA"} since two summary tables are reported. Available are
\itemize{
\item \code{individual}, the default
\item \code{group}, group level results according to Hageman & Arrindell (1999)
}}
}
\description{
Default Method To Get A Summary Table From A cs_analysis Object
}
