% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_get_summary.R
\name{cs_get_summary}
\alias{cs_get_summary}
\title{Get A Summary Table From A cs_analysis Object}
\usage{
cs_get_summary(x, which)
}
\arguments{
\item{x}{An object of class \code{cs_analysis}}

\item{which}{Which level of summary table to return. This is only necessary
for method \code{"HA"} since two summary tables are reported. Available are
\itemize{
\item \code{individual}, the default
\item \code{group}, group level results according to Hageman & Arrindell (1999)
}}
}
\value{
A tibble with clinical significance categories
}
\description{
Retrieve the summary table in a tidy tibble format. This is especially useful
to plot the results or conduct sensitivity analyses.
}
\references{
\itemize{
\item Hageman, W. J., & Arrindell, W. A. (1999). Establishing clinically significant change: increment of precision and the distinction between individual and group level analysis. Behaviour Research and Therapy, 37(12), 1169–1193. https://doi.org/10.1016/S0005-7967(99)00032-7
}
}
