% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_categories.R
\name{.calc_categories_jacobson}
\alias{.calc_categories_jacobson}
\title{Calculate Categories based on an RCI}
\usage{
.calc_categories_jacobson(data, cutoff, rci, rci_cutoff = 1.96, direction = 1)
}
\arguments{
\item{data}{A preprocessed data frame}

\item{cutoff}{A list containing cutoff information}

\item{rci}{A vector containing RCIs}

\item{rci_cutoff}{A numveric value, the multiplier according to which a
patient is considered reliably changed}

\item{direction}{Which direction is beneficial? \code{1} = higher values are
better, \code{-1} = lower values are better}
}
\value{
The full data frame with categories
}
\description{
Calculate Categories based on an RCI
}
