% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_cutoff.R
\name{.calc_cutoff}
\alias{.calc_cutoff}
\title{Calculate Cutoff Criterion Using Summary Values}
\usage{
.calc_cutoff(
  m_clinical,
  sd_clinical,
  m_functional,
  sd_functional,
  type,
  better_is = c("lower", "higher")
)
}
\arguments{
\item{m_clinical}{Mean of clinical population.}

\item{sd_clinical}{SD of clinical population.}

\item{m_functional}{Mean of functional population.}

\item{sd_functional}{SD of functional population.}

\item{type}{Cutoff type}

\item{better_is}{Direction of beneficial results}
}
\description{
Calculate Cutoff Criterion Using Summary Values
}
