% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cutoff.R
\name{.calc_cutoff_data}
\alias{.calc_cutoff_data}
\title{Calculate the Cutoff Criterion}
\usage{
.calc_cutoff_data(
  data,
  m_functional = NA,
  sd_functional = NA,
  type = "a",
  direction = 1
)
}
\arguments{
\item{data}{A preprocessed wide dataframe}

\item{m_functional}{Mean of functional population}

\item{sd_functional}{SD of functional population}

\item{type}{Cutoff type}

\item{direction}{Which direction is beneficial? \code{1} = higher values are
better, \code{-1} = lower values are better}
}
\description{
Calculate the Cutoff Criterion
}
