% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rci.R
\name{.calc_rci_jacobson}
\alias{.calc_rci_jacobson}
\title{RCI for the Jacobson method}
\usage{
.calc_rci_jacobson(data, reliability)
}
\arguments{
\item{data}{A preprocessed dataframe}

\item{reliability}{Instrument's reliability}
}
\value{
A vector with RCIs
}
\description{
This function expects a data frame with at least columns \code{pre} and \code{change}.
The reliability must be a value between 0 and 1.
}
