% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rci.R
\name{.calc_s_diff}
\alias{.calc_s_diff}
\title{Calculate standard error of the instrument}
\usage{
.calc_s_diff(sd_pre, reliability)
}
\arguments{
\item{sd_pre}{Pre standard deviation}

\item{reliability}{Instrument's reliability}
}
\value{
A number
}
\description{
\deqn{S_{diff} = \frac{x_2 - x_1}{S_E}}
\deqn{S_E = SD_{pre} \sqrt{1 - r}}
With \eqn{r} being the reliability
}
