% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_significance.R
\name{.create_summary_table}
\alias{.create_summary_table}
\title{Create Clinical Significance Summary Table}
\usage{
.create_summary_table(x, ...)
}
\arguments{
\item{x}{A clinigsig object}

\item{...}{Additional arguments}
}
\description{
Create Clinical Significance Summary Table
}
