% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plotting_band.R
\name{generate_plotting_band.cs_en}
\alias{generate_plotting_band.cs_en}
\title{Generate RCI Band for EN Method}
\usage{
\method{generate_plotting_band}{cs_en}(x, lower_limit = 0, upper_limit = 100, ...)
}
\description{
Generate RCI Band for EN Method
}
\keyword{internal}
