% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plotting_band.R
\name{generate_plotting_band.cs_gln}
\alias{generate_plotting_band.cs_gln}
\title{Generate RCI Band for GLN Method}
\usage{
\method{generate_plotting_band}{cs_gln}(x, lower_limit = 0, upper_limit = 100, ...)
}
\description{
Generate RCI Band for GLN Method
}
\keyword{internal}
