% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plotting_band.R
\name{generate_plotting_band.cs_jt}
\alias{generate_plotting_band.cs_jt}
\title{Generate RCI Band for JT Method}
\usage{
\method{generate_plotting_band}{cs_jt}(x, lower_limit = 0, upper_limit = 100, ...)
}
\description{
Generate RCI Band for JT Method
}
\keyword{internal}
