% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plotting_band.R
\name{generate_plotting_band.cs_percentage}
\alias{generate_plotting_band.cs_percentage}
\title{Generate RCI Band for the Percentage-Change Approach}
\usage{
\method{generate_plotting_band}{cs_percentage}(x, lower_limit = 0, upper_limit = 100, ...)
}
\description{
Generate RCI Band for the Percentage-Change Approach
}
\keyword{internal}
