% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plotting_band.R
\name{generate_plotting_band}
\alias{generate_plotting_band}
\title{Generic to Calculate RCI Band Data for Plotting}
\usage{
generate_plotting_band(x, lower_limit, upper_limit, ...)
}
\arguments{
\item{x}{An object of respective class}

\item{lower_limit}{Lower plotting limit}

\item{upper_limit}{Upper plotting limit}

\item{...}{Additional arguments}
}
\value{
A tibble
}
\description{
This is an internal generic to generate data for the RCI band in clinical
significance plots. This should not be called directly.
}
