% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_rci_band.R
\name{generate_rci_band.cs_jt}
\alias{generate_rci_band.cs_jt}
\title{Generate RCI Band for JT Method}
\usage{
\method{generate_rci_band}{cs_jt}(x, lower_limit = 0, upper_limit = 100, ...)
}
\arguments{
\item{x}{An object of respective class}

\item{lower_limit}{Lower plotting limit}

\item{upper_limit}{Upper plotting limit}

\item{...}{Additional arguments}
}
\value{
A tibble
}
\description{
Generate RCI Band for JT Method
}
