% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_augmented_data.R
\name{get_augmented_data}
\alias{get_augmented_data}
\title{Get Data Augmented With Clinical Significance Categories}
\usage{
get_augmented_data(x)
}
\arguments{
\item{x}{A clinisig object.}
}
\value{
A tibble with used data and clinical significance categories
}
\description{
Get Data Augmented With Clinical Significance Categories
}
