% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_beneficial_direction.R
\name{get_beneficial_direction}
\alias{get_beneficial_direction}
\title{Get the Specified Beneficial Direction of a Clinical Significance Analysis}
\usage{
get_beneficial_direction(x, ...)
}
\arguments{
\item{x}{A clinisig object}

\item{...}{Additional arguments}
}
\value{
A string
}
\description{
Get the Specified Beneficial Direction of a Clinical Significance Analysis
}
\examples{
results <- jacobson_1989 \%>\%
clinical_significance(
  id = subject,
  time = time,
  outcome = gds,
  pre = "pre",
  reliability = 0.80
)

get_beneficial_direction(results)
}
