% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get Data From A clinisig Object}
\usage{
get_data(x, dataset = "data")
}
\arguments{
\item{x}{A clinisig object.}

\item{dataset}{The dataset you wish to retrieve. Available options are
\code{"original"} (the raw original dataset), \code{"wide"} (the original dataset
in wide format), and \code{"data"} (the dataset which is used in the
calculations). The default is \code{"data"}}
}
\value{
A tibble
}
\description{
Get Data From A clinisig Object
}
