% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_used.R
\name{get_data_used}
\alias{get_data_used}
\title{Retrieve used data from a clinisig object}
\usage{
get_data_used(x, ...)
}
\arguments{
\item{x}{A clinisig object}

\item{...}{Additional arguments}
}
\value{
A tibble with data used in calculations
}
\description{
Retrieve used data from a clinisig object
}
