% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_descriptives.R
\name{get_descriptives}
\alias{get_descriptives}
\title{Get Descriptives Used In The Cutoff Calculation}
\usage{
get_descriptives(x)
}
\arguments{
\item{x}{A clinisg object}
}
\value{
A tibble with means and standard deviations of the clinical and
functional population
}
\description{
Get Descriptives Used In The Cutoff Calculation
}
