% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_method.R
\name{get_method}
\alias{get_method}
\title{Get the employed clinical significance method}
\usage{
get_method(x)
}
\arguments{
\item{x}{A clinisig object}
}
\value{
A string
}
\description{
You may choose between several methods to conduct clinical significance
analyses. This function can retrieve the employed method of a clinisig
object.
}
\examples{
results <- jacobson_1989 \%>\%
clinical_significance(
  id = subject,
  time = time,
  outcome = gds,
  pre = "pre",
  reliability = 0.80
)

get_method(results)
}
