% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n.R
\name{get_n}
\alias{get_n}
\title{Get Number Of Participants From A clinisig Object}
\usage{
get_n(x, which = "all")
}
\arguments{
\item{x}{A clinisg object}

\item{which}{Which n should be returned? Available options are \code{"all"} (n
in the original and used data set), \code{"original"} (n in the original
dataset), and \code{"used"} (n in the used data set, so after conversion to
wide format and omitting cases with missing values). The default is \code{"all"}}
}
\value{
A tibble with number of participants
}
\description{
Get Number Of Participants From A clinisig Object
}
