% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reliability.R
\name{get_reliability}
\alias{get_reliability}
\title{Get Reliability Of A clinisig Object}
\usage{
get_reliability(x)
}
\arguments{
\item{x}{A clinisig object}
}
\value{
A tibble conatining the reliability
}
\description{
Get Reliability Of A clinisig Object
}
