% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary.R
\name{get_summary}
\alias{get_summary}
\title{Get A Summary Table From A clinisig Object}
\usage{
get_summary(x)
}
\arguments{
\item{x}{A clinisig object.}
}
\value{
A tibble with clinical significance categories
}
\description{
Get A Summary Table From A clinisig Object
}
