% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary_table.R
\name{get_summary_table}
\alias{get_summary_table}
\title{Get A Summary Table From A clinisig Object}
\usage{
get_summary_table(x, which = c("individual", "group"))
}
\arguments{
\item{x}{A clinisig object.}

\item{which}{Which level of summary table to return. This is only and issue
for method \code{"HA"} since two summary tables are reported.}
}
\value{
A tibble with clinical significance categories
}
\description{
Get A Summary Table From A clinisig Object
}
