% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{hechler_2014}
\alias{hechler_2014}
\title{Chronic Pain Data}
\format{
A tibble with 208 rows and 3 variables:

\describe{
\item{\code{patient}}{Patient identifier}
\item{\code{measurement}}{Indicator of measurement}
\item{\code{disability}}{Pain-related disability as measured with the PPDI
(lower is better)
}
}
}
\usage{
hechler_2014
}
\description{
A reduced version of the data collected by Hechler et al. (2014)
}
\references{
Hechler, T., Ruhe, A.‑K., Schmidt, P., Hirsch, J., Wager, J.,
Dobe, M., Krummenauer, F., & Zernikow, B. (2014). Inpatient Based Intensive
Interdisciplinary Pain Treatment for Highly Impaired Children with Severe
Chronic Pain: Randomized Controlled Trial of Efficacy and Economic Effects.
Pain, 155(1), 118–128. https://doi.org/10.1016/j.pain.2013.09.015
}
\keyword{datasets}
