% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{jacobson_1989}
\alias{jacobson_1989}
\title{Marital Therapy Data}
\format{
An object of class \code{tbl_df} with 60 rows and 4 columns.

\describe{
\item{subject}{Subject ID}
\item{time}{Measurement}
\item{das}{Dyadic Adjustment Scale score (higher is better)}
\item{gds}{Global Distress Scale score (lower is better)}
}
}
\usage{
jacobson_1989
}
\description{
A dataset containing the data from Jacobson et al. (1989). The purpose of the
study was to examine two forms of behavioral marital therapy,
}
\references{
\itemize{
\item Jacobson, N. S., Schmaling, K. B., Holtzworth-Munroe, A., Katt,
J. L., Wood, L. F., & Follette, V. M. (1989). Research-structured vs
clinically flexible versions of social learning-based marital therapy.
Behaviour Research and Therapy, 27(2), 173-180.
https://doi.org/10.1016/0005-7967(89)90076-4
\item Jacobson, N. S., & Truax, P.
(1991). Clinical significance: A statistical approach to defining meaningful
change in psychotherapy research. Journal of Consulting and Clinical
Psychology, 59(1), 12-19. https://doi.org/10.1037//0022-006X.59.1.12
}
}
\keyword{datasets}
