% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.clinisig}
\alias{plot.clinisig}
\title{Plot Clinical Significance Results}
\usage{
\method{plot}{clinisig}(
  x,
  lower_limit = 0,
  upper_limit = 100,
  rci_fill = "grey10",
  rci_alpha = 0.1,
  ab_line_color = "black",
  show = NULL,
  include_cutoff = TRUE,
  include_cutoff_band = FALSE,
  x_lab = "Pre",
  y_lab = "Post",
  limit_tolerance = 0.02,
  ...
)
}
\arguments{
\item{x}{A clinisig object}

\item{lower_limit}{Numeric, lower plotting limit. Defaults to 0}

\item{upper_limit}{Numeric, upper plotting limit. Defaults to 100}

\item{rci_fill}{String, a color (name or HEX code) for RCI filling}

\item{rci_alpha}{Numeric, controls the transparency of the RCI. This can be
any value between 0 and 1.}

\item{ab_line_color}{String, a color (name or HEX code) for the line
indicating no change.}

\item{show}{String, which category should be shown with distinctive colors?
Available are \code{"recovered"}, \code{"improved"}, \code{"deteriorated"}, or
\code{"unchanged"}.}

\item{include_cutoff}{Logical, should the clinical cutoff be plotted as well?
Defaults to \code{TRUE}.}

\item{include_cutoff_band}{If method was HA, a region of uncertainty around
the cutoff can be plotted}

\item{x_lab}{String, x axis label. Default is \code{"Pre"}.}

\item{y_lab}{String, y axis label. Default is \code{"Post"}.}

\item{limit_tolerance}{Numeric, control amount of overplotting. Defaults to
0.02 (i.e., 2\% of range between lower and upper limit).}

\item{...}{Additional arguments}
}
\value{
A ggplot2 plot
}
\description{
Plot Clinical Significance Results
}
