% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cs_statistical}
\alias{plot.cs_statistical}
\title{Plot an Object of Class cs_statistical}
\usage{
\method{plot}{cs_statistical}(
  x,
  x_lab = "Pre",
  y_lab = "Post",
  color_lab = "Group",
  include_cutoff = TRUE,
  lower_limit,
  upper_limit,
  show,
  point_alpha = 1,
  overplotting = 0.02,
  ...
)
}
\arguments{
\item{x}{An object of class \code{cs_statistical}}

\item{x_lab}{String, x axis label. Default is \code{"Pre"}.}

\item{y_lab}{String, x axis label. Default is \code{"Post"}.}

\item{color_lab}{String, color label (if colors are displayed). Default is
\code{"Group"}}

\item{include_cutoff}{Logical, whether to include the population cutoff.
Default is \code{TRUE}.}

\item{lower_limit}{Numeric, lower plotting limit. Defaults to 2\% smaller
than minimum instrument score}

\item{upper_limit}{Numeric, upper plotting limit. Defaults to 2\% larger than
maximum instrument score}

\item{show}{Unquoted category name. You have several options to color
different features. Available are
\itemize{
\item \code{category} (shows all categories at once)
\item \code{clinical_pre} (shows participants with clinical scores pre intervention)
\item \code{functional_post} (shows participants with functional scores post
intervention)
\item \code{unchanged} (shows unchanged participants)
}}

\item{overplotting}{Numeric, control amount of overplotting. Defaults to 0.02
(i.e., 2\% of range between lower and upper limit).}

\item{...}{Additional arguments}
}
\value{
A ggplot2 plot
}
\description{
Plot an Object of Class cs_statistical
}
\examples{
claus_results <- claus_2020 |>
  cs_statistical(id, time, bdi, m_functional = 8, sd_functional = 7, pre = 1, post = 4, cutoff_type = "c")

plot(claus_results)
plot(claus_results, show = category)


claus_results_ha <- claus_2020 |>
  cs_statistical(id, time, bdi, m_functional = 8, sd_functional = 7, pre = 1, post = 4, reliability = 0.80, cutoff_type = "c", cutoff_method = "HA")

plot(claus_results_ha)
plot(claus_results_ha, show = category)
}
