% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plotting_band.R
\name{generate_plotting_band}
\alias{generate_plotting_band}
\alias{generate_plotting_band.cs_jt}
\alias{generate_plotting_band.cs_gln}
\alias{generate_plotting_band.cs_hll}
\alias{generate_plotting_band.cs_en}
\alias{generate_plotting_band.cs_nk}
\alias{generate_plotting_band.cs_ha}
\alias{generate_plotting_band.cs_percentage}
\alias{generate_plotting_band.cs_anchor_individual_within}
\title{Generic to Calculate RCI Band Data for Plotting}
\usage{
generate_plotting_band(x, lower_limit, upper_limit, ...)

\method{generate_plotting_band}{cs_jt}(x, lower_limit = 0, upper_limit = 100, ...)

\method{generate_plotting_band}{cs_gln}(x, lower_limit = 0, upper_limit = 100, ...)

\method{generate_plotting_band}{cs_hll}(x, lower_limit = 0, upper_limit = 100, ...)

\method{generate_plotting_band}{cs_en}(x, lower_limit = 0, upper_limit = 100, ...)

\method{generate_plotting_band}{cs_nk}(x, lower_limit = 0, upper_limit = 100, ...)

\method{generate_plotting_band}{cs_ha}(x, lower_limit = 0, upper_limit = 100, ...)

\method{generate_plotting_band}{cs_percentage}(x, lower_limit = 0, upper_limit = 100, ...)

\method{generate_plotting_band}{cs_anchor_individual_within}(x, lower_limit = 0, upper_limit = 100, ...)
}
\arguments{
\item{x}{An object of respective class}

\item{lower_limit}{Lower plotting limit}

\item{upper_limit}{Upper plotting limit}

\item{...}{Additional arguments}
}
\value{
A tibble with data for the ribbon indicating unchanged patients
}
\description{
This is an internal generic to generate data for the RCI band in clinical
significance plots. This should not be called directly.
}
