% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_cutoff.R
\name{prep_cutoff}
\alias{prep_cutoff}
\title{Calculate the Cutoff Criterion}
\usage{
prep_cutoff(data, m_functional = NA, sd_functional = NA, type = "a", better_is)
}
\arguments{
\item{data}{A preprocessed wide dataframe}

\item{m_functional}{Mean of functional population}

\item{sd_functional}{SD of functional population}

\item{type}{Cutoff type}

\item{better_is}{Direction of beneficial results}
}
\description{
Calculate the Cutoff Criterion
}
