% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{Prepare Data for Clinical Significance Analyses}
\usage{
prep_data(data, id, time, outcome, measurements = NULL, baseline = NULL)
}
\arguments{
\item{data}{A tidy dataframe}

\item{id}{Participant ID}

\item{time}{Time variable}

\item{outcome}{Outcome variable}

\item{measurements}{If \code{time} contains more than two values, you can specify
your two measurements of interest as a vector}

\item{baseline}{If your \code{time} is of type character, you can specify the pre
measurement}
}
\value{
A wide dataframe containing the ID, pre and post scores, and the
change score
}
\description{
Prepare a tidy dataset for analyses of clinical significance. This function
will select only the relevant variables and spread the measurement column out
to make the data wide which is more optimal for such analyses. Additionally,
\code{prep_data()} calculates the raw change.
}
\details{
Be careful to have your measurement column structured correctly or give
additional information to return correct pre and post columns. If your \code{time}
column contains more than two values, you must choose the two values for
which you would like to calculate the clinical significance (i.e., the pre
and post measurement).

If the \code{time} column is numeric, the values will be sorted numerically. If
the \code{time} column is of type character, you must provide a string for
\code{baseline} (i.e., the pre measurement) because characters cannot be sorted in
a meansingful way.
}
\examples{
prep_data(jacobson_1989, subject, time, das, baseline = "pre")
}
