% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_rci.R
\name{prep_rci}
\alias{prep_rci}
\title{Calculate RCI scores}
\usage{
prep_rci(data, reliability)
}
\arguments{
\item{data}{A preprocessed wide dataframe}

\item{reliability}{The instrument's reliability}
}
\description{
Calculate RCI scores
}
