% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_significance.R
\name{print.clinicsig}
\alias{print.clinicsig}
\title{Print Clinical Significance Results}
\usage{
\method{print}{clinicsig}(x, ...)
}
\arguments{
\item{x}{A clinicsig object}

\item{...}{Additional arguments}
}
\description{
Print Clinical Significance Results
}
