% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_combined.R
\name{print.cs_combined}
\alias{print.cs_combined}
\title{Print Method for the Combined Approach}
\usage{
\method{print}{cs_combined}(x, ...)
}
\arguments{
\item{x}{An object of class \code{cs_combined}}

\item{...}{Additional arguments}
}
\value{
No return value, called for side effects
}
\description{
Print Method for the Combined Approach
}
\examples{
cs_results <- claus_2020 |>
  cs_combined(id, time, hamd, pre = 1, post = 4, reliability = 0.8)
cs_results
}
