% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_distribution.R
\name{print.cs_distribution}
\alias{print.cs_distribution}
\title{Print Method for the Distribution-Based Approach}
\usage{
\method{print}{cs_distribution}(x, ...)
}
\arguments{
\item{x}{An object of class \code{cs_distribution}}

\item{...}{Additional arguments}
}
\value{
No return value, called for side effects
}
\description{
Print Method for the Distribution-Based Approach
}
\examples{
cs_results <- claus_2020 |>
  cs_distribution(id, time, hamd, pre = 1, post = 4, reliability = 0.8)
cs_results
}
