% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_percentage.R
\name{print.cs_percentage_sensitivity}
\alias{print.cs_percentage_sensitivity}
\title{Print Method for the Percentage-Change Approach Sensitivity}
\usage{
\method{print}{cs_percentage_sensitivity}(x, ...)
}
\arguments{
\item{x}{An object of class \code{cs_percentage_sensitivity}}

\item{...}{Additional arguments}
}
\value{
No return value, called for side effects
}
\description{
Print Method for the Percentage-Change Approach Sensitivity
}
