% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_significance.R
\name{summary.clinisig}
\alias{summary.clinisig}
\title{Summary Method for a clinisig object}
\usage{
\method{summary}{clinisig}(object, ...)
}
\arguments{
\item{object}{A clinisig object}

\item{...}{Additional arguments}
}
\description{
Summary Method for a clinisig object
}
