% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_anchor.R
\name{summary.cs_anchor_individual}
\alias{summary.cs_anchor_individual}
\title{Summary Method for the Anchor-Based Approach}
\usage{
\method{summary}{cs_anchor_individual}(object, ...)
}
\arguments{
\item{object}{An object of class \code{cs_anchor_individual}}

\item{...}{Additional arguments}
}
\value{
No return value, called for side effects only
}
\description{
Summary Method for the Anchor-Based Approach
}
\examples{
cs_results <- claus_2020 |>
  cs_anchor(
    id,
    time,
    bdi,
    pre = 1,
    post = 4,
    mid_improvement = 7
  )

cs_results
}
