% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_anchor.R
\name{summary.cs_anchor_individual_sensitivity}
\alias{summary.cs_anchor_individual_sensitivity}
\title{Summary Method for the Anchor-Based Approach Sensitivity}
\usage{
\method{summary}{cs_anchor_individual_sensitivity}(object, ...)
}
\arguments{
\item{object}{An object of class \code{cs_anchor_individual_sensitivity}}

\item{...}{Additional arguments}
}
\value{
No return value, called for side effects only
}
\description{
Summary Method for the Anchor-Based Approach Sensitivity
}
