% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_combined.R
\name{summary.cs_combined}
\alias{summary.cs_combined}
\title{Summary Method for the Combined Approach}
\usage{
\method{summary}{cs_combined}(object, ...)
}
\arguments{
\item{object}{An object of class \code{cs_combined}}

\item{...}{Additional arguments}
}
\value{
No return value, called for side effects only
}
\description{
Summary Method for the Combined Approach
}
\examples{
cs_results <- claus_2020 |>
  cs_combined(id, time, hamd, pre = 1, post = 4, reliability = 0.8)

summary(cs_results)
}
