% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_combined.R
\name{summary.cs_combined_sensitivity}
\alias{summary.cs_combined_sensitivity}
\title{Summary Method for the Combined Approach Sensitivity}
\usage{
\method{summary}{cs_combined_sensitivity}(object, ...)
}
\arguments{
\item{object}{An object of class \code{cs_combined_sensitivity}}

\item{...}{Additional arguments}
}
\value{
No return value, called for side effects only
}
\description{
Summary Method for the Combined Approach Sensitivity
}
