% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_distribution.R
\name{summary.cs_distribution}
\alias{summary.cs_distribution}
\title{Summary Method for the Distribution-Based Approach}
\usage{
\method{summary}{cs_distribution}(object, ...)
}
\arguments{
\item{object}{An object of class \code{cs_distribution}}

\item{...}{Additional arguments}
}
\value{
No return value, called for side effects only
}
\description{
Summary Method for the Distribution-Based Approach
}
\examples{
cs_results <- claus_2020 |>
  cs_distribution(id, time, hamd, pre = 1, post = 4, reliability = 0.8)

summary(cs_results)
}
