% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_distribution.R
\name{summary.cs_distribution_sensitivity}
\alias{summary.cs_distribution_sensitivity}
\title{Summary Method for the Distribution-Based Approach Sensitivity}
\usage{
\method{summary}{cs_distribution_sensitivity}(object, ...)
}
\arguments{
\item{object}{An object of class \code{cs_distribution_sensitivity}}

\item{...}{Additional arguments}
}
\value{
No return value, called for side effects only
}
\description{
Summary Method for the Distribution-Based Approach Sensitivity
}
\examples{
cs_results <- claus_2020 |>
  cs_distribution(
    id, time, bdi, reliability = seq(0.5, 0.9, by = 0.1), pre = 1, post = 4
  )

summary(cs_results)
}
