% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_percentage.R
\name{summary.cs_percentage}
\alias{summary.cs_percentage}
\title{Summary Method for the Percentage-Change Approach}
\usage{
\method{summary}{cs_percentage}(object, ...)
}
\arguments{
\item{object}{An object of class \code{cs_percentage}}

\item{...}{Additional arguments}
}
\value{
No return value, called for side effects only
}
\description{
Summary Method for the Percentage-Change Approach
}
\examples{
cs_results <- claus_2020 |>
  cs_percentage(id, time, hamd, pre = 1, post = 4, pct_improvement = 0.5)

summary(cs_results)
}
