% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_percentage.R
\name{summary.cs_percentage_sensitivity}
\alias{summary.cs_percentage_sensitivity}
\title{Summary Method for the Percentage-Change Approach Sensitivity}
\usage{
\method{summary}{cs_percentage_sensitivity}(object, ...)
}
\arguments{
\item{object}{An object of class \code{cs_percentage_sensitivity}}

\item{...}{Additional arguments}
}
\value{
No return value, called for side effects only
}
\description{
Summary Method for the Percentage-Change Approach Sensitivity
}
\examples{
cs_results <- claus_2020 |>
  cs_percentage(
    id,
    time,
    bdi,
    pre = 1,
    post = 4,
    pct_improvement = seq(0.3, 0.6, by = 0.1)
  )

summary(cs_results)
}
