% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_statistical.R
\name{summary.cs_statistical_sensitivity}
\alias{summary.cs_statistical_sensitivity}
\title{Summary Method for the Statistical Approach Sensitivity}
\usage{
\method{summary}{cs_statistical_sensitivity}(object, ...)
}
\arguments{
\item{object}{An object of class \code{cs_statistical_sensitivity}}

\item{...}{Additional arguments}
}
\value{
No return value, called for side effects only
}
\description{
Summary Method for the Statistical Approach Sensitivity
}
\examples{
cs_results <- claus_2020 |>
  cs_statistical(
    id, time, hamd, pre = 1, post = 4,
    m_functional = seq(7, 9, by = 1),
    sd_functional = 7,
    cutoff_type = "c"
  )

summary(cs_results)
}
