% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_anchor.R
\name{calc_anchor.cs_anchor_group_between}
\alias{calc_anchor.cs_anchor_group_between}
\title{Anchor Calculations for Group Effect Between}
\usage{
\method{calc_anchor}{cs_anchor_group_between}(
  data,
  mid_improvement,
  mid_deterioration,
  reference_group,
  post,
  direction,
  ci_level
)
}
\arguments{
\item{data}{A datasets object}

\item{mid_improvement}{Numeric, change that indicates a clinically
significant improvement}

\item{mid_deterioration}{Numeric, change that indicates a clinically
significant deterioration}

\item{direction}{Which direction is beneficial? Lower = -1, better = 1}

\item{ci_level}{Numeric, desired confidence interval.}
}
\value{
An object of class \code{cs_anchor_group_between}
}
\description{
This is an internal function and should never be called directly.
}
