% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_anchor.R
\name{calc_anchor.cs_anchor_individual}
\alias{calc_anchor.cs_anchor_individual}
\title{Anchor Calculations for Individual Results}
\usage{
\method{calc_anchor}{cs_anchor_individual}(data, mid_improvement, mid_deterioration, direction)
}
\arguments{
\item{data}{A datasets object}

\item{mid_improvement}{Numeric, change that indicates a clinically
significant improvement}

\item{mid_deterioration}{Numeric, change that indicates a clinically
significant deterioration}

\item{direction}{Which direction is beneficial? Lower = -1, better = 1}
}
\value{
An object of class \code{cs_anchor_individual}
}
\description{
This is an internal function and should never be called directly.
}
