% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_anchor.R
\name{calc_anchor}
\alias{calc_anchor}
\title{Generic to Calculate Anchor-Based Results}
\usage{
calc_anchor(data, mid_improvement, mid_deterioration, direction, ci_level)
}
\arguments{
\item{data}{A datasets object}

\item{mid_improvement}{Numeric, change that indicates a clinically
significant improvement}

\item{mid_deterioration}{Numeric, change that indicates a clinically
significant deterioration}

\item{direction}{Which direction is beneficial? Lower = -1, better = 1}
}
\value{
An object of class \code{cs_anchor}
}
\description{
Generic to Calculate Anchor-Based Results
}
