% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rci.R
\name{calc_rci}
\alias{calc_rci}
\title{Generic to Calculate RCI and Associated Change}
\usage{
calc_rci(data, ...)
}
\arguments{
\item{data}{Prepped data with class \verb{cs_*}}

\item{...}{Additional arguments for the specific RCI method}
}
\value{
RCI result object with class \code{cs_distribution}
}
\description{
This is an internal generic and should not be called directly. Depending on
the different RCI method requested by the user, the appropriate method is
called. It calculates the RCI and according clinical significance category
for each participant.
}
\keyword{internal}
