% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rci.R
\name{calc_rci}
\alias{calc_rci}
\title{Generic to Calculate RCI and Associated Change}
\usage{
calc_rci(
  data,
  m_pre,
  m_post,
  sd_pre,
  sd_post,
  reliability,
  reliability_post,
  direction,
  critical_value
)
}
\arguments{
\item{data}{Prepped data with class \verb{cs_*}}

\item{m_pre}{Pre measurement mean}

\item{m_post}{Post measurement mean}

\item{sd_pre}{Pre measurement SD}

\item{sd_post}{Post measurement SD}

\item{reliability}{Instrument reliability}

\item{reliability_post}{Instrument reliability at post measurement}

\item{direction}{Beneficial intervention effect for given instrument. 1 =
higher is better, -1 = lower is better}

\item{critical_value}{Critical RCI value, typically 1.96}
}
\value{
RCI result object with class \code{cs_distribution}
}
\description{
This is an internal generic and should not be called directly. Depending on
the different RCI method requested by the user, the appropriate method is
called.
}
