% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_significance.R
\name{clinical_significance}
\alias{clinical_significance}
\title{Clinical Significance}
\usage{
clinical_significance(
  data,
  id,
  time,
  outcome,
  pre = NULL,
  post = NULL,
  m_functional = NA,
  sd_functional = NA,
  type = "a",
  reliability,
  better_is = c("lower", "higher"),
  method = c("JT", "GLN", "EN")
)
}
\arguments{
\item{data}{A tidy data frame}

\item{id}{Participant ID}

\item{time}{Time variable}

\item{outcome}{Outcome variable}

\item{pre}{Pre measurement}

\item{post}{Post measurement}

\item{m_functional}{Mean of the functional population}

\item{sd_functional}{Standard deviation of the functional population}

\item{type}{Cutoff type. Available are \code{"a"}, \code{"b"}, and \code{"c"}. Defaults to
\code{"a"} (see details for further information in which cutoff to choose).}

\item{reliability}{The instrument's reliability estimate.}

\item{better_is}{Which direction means a better outcome? Available are
\code{"lower"} and \code{"higher"}. Defaults to \code{"higher"}.}

\item{method}{Clinical significance method. Available are \verb{\\"JT\\"} (Jacobson
& Truax) and \verb{\\"GLN\\"} (Gulliksen, Lord, and Novick)}
}
\value{
A S3 object of class \code{clinisig}
}
\description{
Clinical Significance
}
