% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_anchor_group.R
\name{cs_anchor_group}
\alias{cs_anchor_group}
\alias{cs_anchor_group.default}
\title{Groupwise Anchor-Based Analysis of Clinical Significance}
\usage{
cs_anchor_group(data, ...)

\method{cs_anchor_group}{default}(
  data,
  id,
  time,
  outcome,
  group = NULL,
  pre = NULL,
  post = NULL,
  mid_improvement = NULL,
  mid_deterioration = NULL,
  better_is = c("lower", "higher"),
  effect = c("within", "between"),
  bayesian = TRUE,
  prior_scale = "medium",
  reference_group = NULL,
  ci_level = 0.95,
  ...
)
}
\arguments{
\item{data}{A tidy data frame}

\item{...}{Additional arguments passed to methods.}

\item{id}{Participant ID}

\item{time}{Time variable}

\item{outcome}{Outcome variable}

\item{group}{Grouping variable (optional)}

\item{pre}{Pre measurement (only needed if the time variable contains more
than two measurements)}

\item{post}{Post measurement (only needed if the time variable contains more
than two measurements)}

\item{mid_improvement}{Numeric, change that indicates a clinically
significant improvement}

\item{mid_deterioration}{Numeric, change that indicates a clinically
significant deterioration (optional). If \code{mid_deterioration} is not
provided, it will be assumed to be equal to \code{mid_improvement}}

\item{better_is}{Which direction means a better outcome for the used
instrument? Available are
\itemize{
\item \code{"lower"} (lower outcome scores are desirable, the default) and
\item \code{"higher"} (higher outcome scores are desirable)
}}

\item{effect}{String, specify which effect should be
calculated. Available are
\itemize{
\item \code{"within"} (the default), which yields the mean pre-post intervention
difference with associated confidence intervals
\item \code{"between"}, which estimates the group wise mean difference and
confidence intervals between two or more groups specified with the \code{group}
argument at the specified measurement supplied with the \code{post}- argument
The reference group may be supplied with \code{reference_group}
}}

\item{bayesian}{Logical. Indicates if a
Bayesian estimate (i.e., the median) of group differences with a credible
interval should be calculated (if set to \code{TRUE}, the default) or a
frequentist mean difference with confidence interval (if set to \code{FALSE})}

\item{prior_scale}{String or numeric, can be adjusted to change the Bayesian
prior distribution. See the documentation for \code{rscale} in
\code{\link[BayesFactor:ttestBF]{BayesFactor::ttestBF()}} for details.}

\item{reference_group}{Specify the reference group to which all subsequent
groups are compared against if \code{effect = "within"}
(optional). Otherwise, the first distinct group is chosen based on
alphabetical, numerical or factor ordering.}

\item{ci_level}{Numeric, define the credible or confidence interval level.
The default is 0.95 for a 95\%-CI.}
}
\value{
An S3 object of class \code{cs_analysis} and \code{cs_anchor}
}
\description{
Groupwise Anchor-Based Analysis of Clinical Significance
}
\section{Methods (by class)}{
\itemize{
\item \code{cs_anchor_group(default)}: Default method for data frames

}}
\references{
Hespanhol, L., Vallio, C. S., Costa, L. M., & Saragiotto, B. T.
(2019). Understanding and interpreting confidence and credible intervals
around effect estimates. Brazilian Journal of Physical Therapy, 23(4),
290–301. https://doi.org/10.1016/j.bjpt.2018.12.006
}
\seealso{
Main clinical signficance functions
\code{\link{cs_anchor}()},
\code{\link{cs_combined}()},
\code{\link{cs_distribution}()},
\code{\link{cs_percentage}()},
\code{\link{cs_statistical}()}
}
\concept{main}
