% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_combined.R
\name{cs_combined}
\alias{cs_combined}
\title{Combined Analysis of Clinical Significance}
\usage{
cs_combined(
  data,
  id,
  time,
  outcome,
  group = NULL,
  pre = NULL,
  post = NULL,
  reliability = NULL,
  reliability_post = NULL,
  m_functional = NULL,
  sd_functional = NULL,
  better_is = c("lower", "higher"),
  rci_method = c("JT", "GLN", "HLL", "EN", "NK", "HA", "HLM"),
  cutoff_type = c("a", "b", "c"),
  significance_level = 0.05
)
}
\arguments{
\item{data}{A tidy data frame}

\item{id}{Participant ID}

\item{time}{Time variable}

\item{outcome}{Outcome variable}

\item{group}{Grouping variable (optional)}

\item{pre}{Pre measurement (only needed if the time variable contains more
than two measurements)}

\item{post}{Post measurement (only needed if the time variable contains more
than two measurements)}

\item{reliability}{The instrument's reliability estimate. If you selected the
NK method, the here specified reliability will be the instrument's pre
measurement reliability}

\item{reliability_post}{The instrument's reliability at post measurement
(only needed for the NK method)}

\item{m_functional}{Numeric, mean of functional population.}

\item{sd_functional}{Numeric, standard deviation of functional population}

\item{better_is}{Which direction means a better outcome for the used
instrument? Available are
\itemize{
\item \code{"lower"} (lower outcome scores are desirable, the default) and
\item \code{"higher"} (higher outcome scores are desirable)
}}

\item{rci_method}{Clinical significance method. Available are
\itemize{
\item \code{"JT"} (Jacobson & Truax, 1991, the default)
\item \code{"GLN"} (Gulliksen, Lord, and Novick; Hsu, 1989, Hsu, 1995)
\item \code{"HLL"} (Hsu, Linn & Nord; Hsu, 1989)
\item \code{"EN"} (Edwards & Nunnally; Speer, 1992)
\item \code{"NK"} (Nunnally & Kotsch, 1983), requires a reliability estimate at post
measurement. If this is not supplied, reliability and reliability_post are
assumed to be equal
\item \code{"HA"} (Hageman & Arrindell, 1999)
\item \code{"HLM"} (Hierarchical Linear Modeling; Raudenbush & Bryk, 2002),
requires at least three measurements per patient
}}

\item{cutoff_type}{Cutoff type. Available are \code{"a"}, \code{"b"}, and \code{"c"}.
Defaults to \code{"a"} but \code{"c"} is usually recommended. For \code{"b"} and \code{"c"},
summary data from a functional population must be given with arguments
\code{"m_functional"} and \code{"sd_functional"}.}

\item{significance_level}{Significance level alpha, defaults to \code{0.05}. If
you choose the \code{"HA"} method, this value corresponds to the maximum risk of
misclassification}
}
\value{
An S3 object of class \code{cs_analysis} and \code{cs_combined}
}
\description{
Combined Analysis of Clinical Significance
}
\examples{
cs_results <- claus_2020 |>
  cs_combined(id, time, bdi, pre = 1, post = 4, reliability = 0.80)

cs_results
summary(cs_results)
plot(cs_results)


# Different cutoff type
cs_results_c <- claus_2020 |>
  cs_combined(id, time, bdi, pre = 1, post = 4, reliability = 0.80, m_functional = 8, sd_functional = 8, cutoff_type = "c")

cs_results_c
summary(cs_results_c)
plot(cs_results_c)


# Grouped analysis
cs_results_grouped <- claus_2020 |>
  cs_combined(id, time, bdi, pre = 1, post = 4, group = treatment, reliability = 0.80, m_functional = 8, sd_functional = 8, cutoff_type = "c")

cs_results_grouped
summary(cs_results_grouped)
plot(cs_results_grouped)
}
\seealso{
Main clinical signficance functions
\code{\link{cs_anchor}()},
\code{\link{cs_distribution}()},
\code{\link{cs_percentage}()},
\code{\link{cs_statistical}()}
}
\concept{main}
