% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_distribution.R
\name{cs_distribution}
\alias{cs_distribution}
\title{Distribution-Based Analysis of Clinical Significance}
\usage{
cs_distribution(
  data,
  id,
  time,
  outcome,
  group = NULL,
  pre = NULL,
  post = NULL,
  reliability = NULL,
  reliability_post = NULL,
  better_is = c("lower", "higher"),
  rci_method = c("JT", "GLN", "HLL", "EN", "NK", "HA", "HLM"),
  significance_level = 0.05
)
}
\arguments{
\item{data}{A tidy data frame}

\item{id}{Participant ID}

\item{time}{Time variable}

\item{outcome}{Outcome variable}

\item{group}{Grouping variable (optional)}

\item{pre}{Pre measurement (only needed if the time variable contains more
than two measurements)}

\item{post}{Post measurement (only needed if the time variable contains more
than two measurements)}

\item{reliability}{The instrument's reliability estimate. If you selected the
NK method, the here specified reliability will be the instrument's pre
measurement reliability}

\item{reliability_post}{The instrument's reliability at post measurement
(only needed for the NK method)}

\item{better_is}{Which direction means a better outcome for the used
instrument? Available are
\itemize{
\item \code{"lower"} (lower outcome scores are desirable, the default) and
\item \code{"higher"} (higher outcome scores are desirable)
}}

\item{rci_method}{Clinical significance method. Available are
\itemize{
\item \code{"JT"} (Jacobson & Truax, 1991, the default)
\item \code{"GLN"} (Gulliksen, Lord, and Novick; Hsu, 1989, Hsu, 1995)
\item \code{"HLL"} (Hsu, Linn & Nord; Hsu, 1989)
\item \code{"EN"} (Edwards & Nunnally; Speer, 1992)
\item \code{"NK"} (Nunnally & Kotsch, 1983), requires a reliability estimate at post
measurement. If this is not supplied, reliability and reliability_post are
assumed to be equal
\item \code{"HA"} (Hageman & Arrindell, 1999)
\item \code{"HLM"} (Hierarchical Linear Modeling; Raudenbush & Bryk, 2002),
requires at least three measurements per patient
}}

\item{significance_level}{Significance level alpha, defaults to \code{0.05}. If
you choose the \code{"HA"} method, this value corresponds to the maximum risk of
misclassification}
}
\value{
An S3 object of class \code{cs_analysis} and \code{cs_distribution}
}
\description{
Distribution-Based Analysis of Clinical Significance
}
\section{Data preparation}{
 The data set must be tidy, which corresponds to a
long data frame in general. It must contain a patient identifier which must
be unique per patient. Also, a column containing the different measurements
and the outcome must be supplied. Each participant-measurement combination
must be unique, so for instance, the data must not contain two "After"
measurements for the same patient.

Additionally, if the measurement column contains only two values, the first
value based on alphabetical, numerical or factor ordering will be used as
the \code{pre} measurement. For instance, if the column contains the
measurements identifiers \code{"pre"} and \code{"post"} as strings, then \code{"post"}
will be sorted before \code{"pre"} and thus be used as the \code{"pre"} measurement.
The function will throw a warning but generally you may want to explicitly
define the \code{"pre"} and \code{"post"} measurement with arguments \code{pre} and
\code{post}. In case of more than two measurement identifiers, you have to
define \code{pre} and \code{post} manually since the function does not know what your
pre and post intervention measurements are.

If your data is grouped, you can specify the group by referencing the
grouping variable (see examples below). The analysis is then run for every
group to compare group differences.
}

\examples{
cs_results <- claus_2020 |>
  cs_distribution(id, time, hamd, pre = 1, post = 4, reliability = 0.80)

cs_results
summary(cs_results)
plot(cs_results)


# Different RCI method
cs_results_ha <- claus_2020 |>
  cs_distribution(id, time, hamd, pre = 1, post = 4, reliability = 0.80, rci_method = "HA")

cs_results_ha
summary(cs_results_ha)
plot(cs_results_ha)


# Grouped analysis
cs_results_grouped <- claus_2020 |>
  cs_distribution(id, time, hamd, pre = 1, post = 4, group = treatment, reliability = 0.80)

cs_results_grouped
summary(cs_results_grouped)
plot(cs_results_grouped)
}
\references{
\itemize{
\item Jacobson, N. S., & Truax, P. (1991). Clinical significance: A statistical approach to defining meaningful change in psychotherapy research. Journal of Consulting and Clinical Psychology, 59(1), 12–19. https://doi.org/10.1037//0022-006X.59.1.12
\item Hsu, L. M. (1989). Reliable changes in psychotherapy: Taking into account regression toward the mean. Behavioral Assessment, 11(4), 459–467.
\item Hsu, L. M. (1995). Regression toward the mean associated with measurement error and the identification of improvement and deterioration in psychotherapy. Journal of Consulting and Clinical Psychology, 63(1), 141–144. https://doi.org/10.1037//0022-006x.63.1.141
\item Speer, D. C. (1992). Clinically significant change: Jacobson and Truax (1991) revisited. Journal of Consulting and Clinical Psychology, 60(3), 402–408. https://doi.org/10.1037/0022-006X.60.3.402
\item Nunnally, J. C., & Kotsch, W. E. (1983). Studies of individual subjects: Logic and methods of analysis. British Journal of Clinical Psychology, 22(2), 83–93. https://doi.org/10.1111/j.2044-8260.1983.tb00582.x
\item Hageman, W. J., & Arrindell, W. A. (1999). Establishing clinically significant change: increment of precision and the distinction between individual and group level analysis. Behaviour Research and Therapy, 37(12), 1169–1193. https://doi.org/10.1016/S0005-7967(99)00032-7
\item Raudenbush, S. W., & Bryk, A. S. (2002). Hierarchical Linear Models - Applications and Data Analysis Methods (2nd ed.). Sage Publications.
}
}
\seealso{
Main clinical signficance functions
\code{\link{cs_anchor}()},
\code{\link{cs_combined}()},
\code{\link{cs_percentage}()},
\code{\link{cs_statistical}()}
}
\concept{main}
