% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_get_n.R
\name{cs_get_n}
\alias{cs_get_n}
\title{Get Number Of Participants From A cs_analysis Object}
\usage{
cs_get_n(x, which = "all")
}
\arguments{
\item{x}{A cs_analysis object}

\item{which}{Which n should be returned? Available options are
\itemize{
\item \code{"all"}, n in the original and used data set (the default)
\item \code{"original"}, n in the original dataset
\item \code{"used"}, n in the used data set, so after conversion to wide format and
omitting cases with missing values
}}
}
\value{
A tibble with number of participants
}
\description{
Get Number Of Participants From A cs_analysis Object
}
\examples{
cs_results <- claus_2020 |>
  cs_anchor(id, time, bdi, mid_improvement = 9, pre = 1, post = 4)

cs_get_n(cs_results)
cs_get_n(cs_results, "original")
cs_get_n(cs_results, "used")
}
