% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_get_reliability.R
\name{cs_get_reliability}
\alias{cs_get_reliability}
\title{Get Reliability Of A cs_analysis Object}
\usage{
cs_get_reliability(x)
}
\arguments{
\item{x}{A cs_analysis object}
}
\value{
A tibble showing the reliability
}
\description{
Get Reliability Of A cs_analysis Object
}
\examples{
cs_results <- claus_2020 |>
  cs_distribution(
    id,
    time,
    bdi,
    pre = 1,
    post = 4,
    reliability = 0.80
  )


cs_get_reliability(cs_results)
}
\seealso{
Extractor functions
\code{\link{cs_get_augmented_data}()},
\code{\link{cs_get_data}()},
\code{\link{cs_get_n}()},
\code{\link{cs_get_summary}()}
}
\concept{get}
