% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_percentage.R
\name{cs_percentage}
\alias{cs_percentage}
\title{Percentage-Change Analysis of Clinical Significance}
\usage{
cs_percentage(
  data,
  id,
  time,
  outcome,
  group = NULL,
  pre = NULL,
  post = NULL,
  pct_improvement = NULL,
  pct_deterioration = NULL,
  better_is = c("lower", "higher")
)
}
\arguments{
\item{data}{A tidy data frame}

\item{id}{Participant ID}

\item{time}{Time variable}

\item{outcome}{Outcome variable}

\item{group}{Grouping variable (optional)}

\item{pre}{Pre measurement (only needed if the time variable contains more
than two measurements)}

\item{post}{Post measurement (only needed if the time variable contains more
than two measurements)}

\item{pct_improvement}{Numeric, percent change that indicates a clinically
significant improvement}

\item{pct_deterioration}{Numeric, percent change that indicates a clinically
significant deterioration}

\item{better_is}{Which direction means a better outcome for the used
instrument? Available are
\itemize{
\item \code{"lower"} (lower outcome scores are desirable, the default) and
\item \code{"higher"} (higher outcome scores are desirable)
}}
}
\value{
An S3 object of class \code{cs_analysis} and \code{cs_percentage}
}
\description{
Percentage-Change Analysis of Clinical Significance
}
\examples{
claus_2020 |>
  cs_percentage(id, time, hamd, pre = 1, post = 4, pct_improvement = 0.5)


# Different thresholds for improvement and deterioration
claus_2020 |>
  cs_percentage(id, time, hamd, pre = 1, post = 4, pct_improvement = 0.5, pct_deterioration = 0.3)
}
