% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_statistical.R
\name{cs_statistical}
\alias{cs_statistical}
\title{Statistical Analysis of Clinical Significance}
\usage{
cs_statistical(
  data,
  id,
  time,
  outcome,
  group = NULL,
  pre = NULL,
  post = NULL,
  m_functional = NULL,
  sd_functional = NULL,
  reliability = NULL,
  better_is = c("lower", "higher"),
  cutoff_method = c("JT", "HA"),
  cutoff_type = c("a", "b", "c"),
  significance_level = 0.05
)
}
\arguments{
\item{data}{A tidy data frame}

\item{id}{Participant ID}

\item{time}{Time variable}

\item{outcome}{Outcome variable}

\item{group}{Grouping variable (optional)}

\item{pre}{Pre measurement (only needed if the time variable contains more
than two measurements)}

\item{post}{Post measurement (only needed if the time variable contains more
than two measurements)}

\item{m_functional}{Numeric, mean of functional population.}

\item{sd_functional}{Numeric, standard deviation of functional population}

\item{reliability}{The instrument's reliability estimate. If you selected
the NK method, the here specified reliability will be the instrument's pre
measurement reliability}

\item{better_is}{Which direction means a better outcome for the used
instrument? Available are
\itemize{
\item \code{"lower"} (lower outcome scores are desirable, the default) and
\item \code{"higher"} (higher outcome scores are desirable)
}}

\item{cutoff_method}{Cutoff method, Available are
\itemize{
\item \code{"JT"} (Jacobson & Truax, 1991, the default)
\item \code{"HA"} (Hageman & Arrindell, 1999)
}}

\item{cutoff_type}{Cutoff type. Available are \code{"a"}, \code{"b"}, and \code{"c"}.
Defaults to \code{"a"} but \code{"c"} is usually recommended. For \code{"b"} and \code{"c"},
summary data from a functional population must be given with arguments
\code{"m_functional"} and \code{"sd_functional"}.}

\item{significance_level}{Significance level alpha, defaults to \code{0.05}. If
you choose the \code{"HA"} method, this value corresponds to the maximum risk
of misclassification}
}
\value{
An S3 object of class \code{clinisig} and \code{cs_statistical}
}
\description{
Statistical Analysis of Clinical Significance
}
\examples{
claus_2020 |>
  cs_statistical(id, time, bdi, pre = 1, post = 4)

# Different cutoff
claus_2020 |>
  cs_statistical(id, time, bdi, m_functional = 8, sd_functional = 7, pre = 1, post = 4, cutoff_type = "c")

# Different method
claus_2020 |>
  cs_statistical(id, time, bdi, m_functional = 8, sd_functional = 7, pre = 1, post = 4, cutoff_type = "c", cutoff_method = "HA")
}
\references{
\itemize{
\item Jacobson, N. S., & Truax, P. (1991). Clinical significance: A statistical approach to defining meaningful change in psychotherapy research. Journal of Consulting and Clinical Psychology, 59(1), 12–19. https://doi.org/10.1037//0022-006X.59.1.12
\item Hageman, W. J., & Arrindell, W. A. (1999). Establishing clinically significant change: increment of precision and the distinction between individual and group level analysis. Behaviour Research and Therapy, 37(12), 1169–1193. https://doi.org/10.1016/S0005-7967(99)00032-7
}
}
