% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cutoff.R
\name{.calc_cutoff}
\alias{.calc_cutoff}
\title{Calculate Cutoff Criterion Using Summary Values}
\usage{
.calc_cutoff(
  m_clinical,
  sd_clinical,
  m_functional,
  sd_functional,
  type = "a",
  direction = 1
)
}
\arguments{
\item{m_clinical}{Mean of clinical population.}

\item{sd_clinical}{SD of clinical population.}

\item{m_functional}{Mean of functional population.}

\item{sd_functional}{SD of functional population.}

\item{type}{Cutoff type}

\item{direction}{Which direction is beneficial? \code{1} = higher values are
better, \code{-1} = lower values are better}
}
\description{
Calculate Cutoff Criterion Using Summary Values
}
