% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_rci.R
\name{.calc_s_diff}
\alias{.calc_s_diff}
\title{Calculate standard error of the instrument}
\usage{
.calc_s_diff(sd_pre, reliability)
}
\arguments{
\item{sd_pre}{Pre standard deviation}

\item{reliability}{Instrument's reliability}
}
\description{
Calculate standard error of the instrument
}
