% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plotting_band.R
\name{generate_plotting_band.cs_anchor_individual}
\alias{generate_plotting_band.cs_anchor_individual}
\title{Generate RCI Band for the Individual Anchor-Based Approach}
\usage{
\method{generate_plotting_band}{cs_anchor_individual}(x, lower_limit = 0, upper_limit = 100, ...)
}
\arguments{
\item{x}{An object of respective class}

\item{lower_limit}{Lower plotting limit}

\item{upper_limit}{Upper plotting limit}

\item{...}{Additional arguments}
}
\value{
A tibble
}
\description{
Generate RCI Band for the Individual Anchor-Based Approach
}
