% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plotting_band.R
\name{generate_plotting_band.cs_hll}
\alias{generate_plotting_band.cs_hll}
\title{Generate RCI Band for HLL Method}
\usage{
\method{generate_plotting_band}{cs_hll}(x, lower_limit = 0, upper_limit = 100, ...)
}
\description{
Generate RCI Band for HLL Method
}
\keyword{internal}
