% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cutoff_descriptives.R
\name{get_cutoff_descriptives}
\alias{get_cutoff_descriptives}
\title{Get Descriptives Used In The Cutoff Calculation}
\usage{
get_cutoff_descriptives(x)
}
\arguments{
\item{x}{A clinisg object}
}
\value{
A tibble with means and standard deviations of the clinical and
functional population
}
\description{
Get Descriptives Used In The Cutoff Calculation
}
\examples{
results <- jacobson_1989 \%>\%
  clinical_significance(
    id = subject,
    time = time,
    outcome = gds,
    pre = "pre",
    reliability = 0.80,
    m_functional = 30,
    sd_functional = 10,
    type = "c"
  )

get_cutoff_descriptives(results)
}
