% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cs_anchor_group_within}
\alias{plot.cs_anchor_group_within}
\title{Plot an Object of Class cs_anchor_group_within}
\usage{
\method{plot}{cs_anchor_group_within}(
  x,
  x_lab = "Group",
  y_lab = "Mean Intervention Effect\\n(with 95\%-CI)",
  ...
)
}
\arguments{
\item{x}{An object of class \code{cs_anchor_group_within}}

\item{x_lab}{String, x axis label. Defaults to \code{"Group"}}

\item{y_lab}{String, y axis label, defaults to
\code{"Mean Intervention Effect (with 95\%-CI)"}}

\item{...}{Additional arguments}
}
\value{
A ggplot2 plot
}
\description{
This function creates a generic clinical significance plot bz plotting the
patients' pre intervention value on the x-axis and the post intervention
score on the y-axis. Additionally, the RCI (region signifying unchanged
patients) is shown with a diagonal corresponding to no change.
}
