% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cs_distribution}
\alias{plot.cs_distribution}
\title{Plot an Object of Class cs_distribution}
\usage{
\method{plot}{cs_distribution}(x, ...)
}
\arguments{
\item{...}{}
}
\description{
Plot an Object of Class cs_distribution
}
