% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_significance.R
\name{print.clinisig}
\alias{print.clinisig}
\title{Print Clinical Significance Results}
\usage{
\method{print}{clinisig}(x, ...)
}
\arguments{
\item{x}{A clinisig object}

\item{...}{Additional arguments passed to \code{export_table()}}
}
\value{
No return value, called for side effects
}
\description{
Print Clinical Significance Results
}
