% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_anchor_group.R
\name{print.cs_anchor_group_between}
\alias{print.cs_anchor_group_between}
\title{Print Method for the Anchor-Based Approach for Groups (Between)}
\usage{
\method{print}{cs_anchor_group_between}(x, ...)
}
\arguments{
\item{x}{An object of class \code{cs_anchor_group_between}}

\item{...}{Additional arguments}
}
\value{
No return value, called for side effects
}
\description{
Print Method for the Anchor-Based Approach for Groups (Between)
}
\examples{
cs_results <- claus_2020 |>
  cs_anchor(
    id,
    time,
    bdi,
    post = 4,
    mid_improvement = 7,
    group = treatment,
    target = "group",
    effect = "between"
  )

cs_results
}
