% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_anchor_group.R
\name{print.cs_anchor_group_within}
\alias{print.cs_anchor_group_within}
\title{Print Method for the Anchor-Based Approach for Groups (Within)}
\usage{
\method{print}{cs_anchor_group_within}(x, ...)
}
\arguments{
\item{x}{An object of class \code{cs_anchor_group_within}}

\item{...}{Additional arguments}
}
\value{
No return value, called for side effects
}
\description{
Print Method for the Anchor-Based Approach for Groups (Within)
}
\examples{
cs_results <- claus_2020 |>
  cs_anchor(
    id,
    time,
    bdi,
    pre = 1,
    post = 4,
    mid_improvement = 7,
    target = "group"
  )

cs_results
}
