% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_anchor.R
\name{print.cs_anchor_individual}
\alias{print.cs_anchor_individual}
\title{Print Method for the Percentange-Change Approach}
\usage{
\method{print}{cs_anchor_individual}(x, ...)
}
\arguments{
\item{x}{An object of class \code{cs_anchor_individual}}

\item{...}{Additional arguments}
}
\value{
No return value, called for side effects
}
\description{
Print Method for the Percentange-Change Approach
}
\examples{
cs_results <- claus_2020 |>
  cs_distribution(id, time, hamd, pre = 1, post = 4, reliability = 0.8)
cs_results
}
