% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_distribution.R
\name{print.cs_distribution_sensitivity}
\alias{print.cs_distribution_sensitivity}
\title{Print Method for the Distribution-Based Approach Sensitivity}
\usage{
\method{print}{cs_distribution_sensitivity}(x, ...)
}
\arguments{
\item{x}{An object of class \code{cs_distribution_sensitivity}}

\item{...}{Additional arguments}
}
\value{
No return value, called for side effects
}
\description{
Print Method for the Distribution-Based Approach Sensitivity
}
