% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_percentage.R
\name{print.cs_percentage}
\alias{print.cs_percentage}
\title{Print Method for the Percentange-Change Approach}
\usage{
\method{print}{cs_percentage}(x, ...)
}
\arguments{
\item{x}{An object of class \code{cs_percentage}}

\item{...}{Additional arguments}
}
\value{
No return value, called for side effects
}
\description{
Print Method for the Percentange-Change Approach
}
\examples{
cs_results <- claus_2020 |>
  cs_percentage(
    id,
    time,
    bdi,
    pre = 1,
    post = 4,
    pct_improvement = 0.5
  )

cs_results
}
