% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_anchor_group.R
\name{summary.cs_anchor_group_between}
\alias{summary.cs_anchor_group_between}
\title{Summary Method for the Anchor-Based Approach for Groups (Between)}
\usage{
\method{summary}{cs_anchor_group_between}(object, ...)
}
\arguments{
\item{object}{An object of class \code{cs_anchor_group_between}}

\item{...}{Additional arguments}
}
\value{
No return value, called for side effects only
}
\description{
Summary Method for the Anchor-Based Approach for Groups (Between)
}
