% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_anchor.R
\name{summary.cs_anchor_group_between}
\alias{summary.cs_anchor_group_between}
\title{Summary Method for the Anchor-Based Approach for Groups (Between)}
\usage{
\method{summary}{cs_anchor_group_between}(object, ...)
}
\arguments{
\item{object}{An object of class \code{cs_anchor_group_between}}

\item{...}{Additional arguments}
}
\value{
No return value, called for side effects only
}
\description{
Summary Method for the Anchor-Based Approach for Groups (Between)
}
\examples{
cs_results <- antidepressants |>
  cs_anchor(
    patient,
    measurement,
    post = "After",
    mom_di,
    mid_improvement = 8,
    target = "group",
    effect = "between",
    group = condition
  )

summary(cs_results)
}
