% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_statistical.R
\name{summary.cs_statistical}
\alias{summary.cs_statistical}
\title{Summary Method for the Statistical Approach}
\usage{
\method{summary}{cs_statistical}(object, ...)
}
\arguments{
\item{object}{An object of class \code{cs_distribution}}

\item{...}{Additional arguments}
}
\value{
No return value, called for side effects only
}
\description{
Summary Method for the Statistical Approach
}
\examples{
cs_results <- claus_2020 |>
  cs_statistical(
    id,
    time,
    hamd,
    pre = 1,
    post = 4,
    m_functional = 8,
    sd_functional = 7
  )

summary(cs_results)
}
